// src/playground/generated/assistActions.ts
// This file is generated by codegen/src/metadata.rs
export const ASSIST_ACTIONS = {
	preset: {
		recommended: "recommended",
		all: "all",
		none: "none"
	},
  source: {
    noDuplicateClasses: "noDuplicateClasses",
    organizeImports: "organizeImports",
    useSortedAttributes: "useSortedAttributes",
    useSortedEnumMembers: "useSortedEnumMembers",
    useSortedInterfaceMembers: "useSortedInterfaceMembers",
    useSortedKeys: "useSortedKeys",
    useSortedPackageJson: "useSortedPackageJson",
    useSortedProperties: "useSortedProperties",
    useSortedSelectionSet: "useSortedSelectionSet",
    useSortedTypeFields: "useSortedTypeFields",
  } as const,
} as const;